add error check

+++ src/main/scala/com/broadinstitute/dsp/resourcevalidator/ErroredRuntimeChecker.scala
@@ -8,38 +8,41 @@ import org.broadinstitute.dsde.workbench.google2.{DataprocClusterName, InstanceN
 import org.broadinstitute.dsde.workbench.model.TraceId
 
 // Interpreter
-object DeletedRuntimeChecker {
-  implicit def apply[F[_]](implicit ev: RuntimeChecker[F]): RuntimeChecker[F] = ev
-
+object ErroredRuntimeChecker {
   def iml[F[_]: Timer](
     deps: RuntimeCheckerDeps[F]
   )(implicit F: Concurrent[F], logger: Logger[F], ev: ApplicativeAsk[F, TraceId]): RuntimeChecker[F] =
     new RuntimeChecker[F] {
-      override def checkType = "deleted-runtime"
-
+      override def checkType = "error-ed-runtime"
       override def dependencies: RuntimeCheckerDeps[F] = deps
 
-      def checkRuntimeStatus(runtime: Runtime,
-                             isDryRun: Boolean)(implicit ev: ApplicativeAsk[F, TraceId]): F[Option[Runtime]] =
-        runtime.cloudService match {
-          case CloudService.Dataproc =>
-            checkClusterStatus(runtime, isDryRun)
-          case CloudService.Gce =>
-            checkGceRuntimeStatus(runtime, isDryRun)
-        }
+      override def checkRuntimeStatus(runtime: Runtime, isDryRun: Boolean)(
+        implicit ev: ApplicativeAsk[F, TraceId]
+      ): F[Option[Runtime]] = runtime.cloudService match {
+        case CloudService.Dataproc =>
+          checkClusterStatus(runtime, isDryRun)
+        case CloudService.Gce =>
+          checkGceRuntimeStatus(runtime, isDryRun)
+      }
 
       def checkClusterStatus(runtime: Runtime,
                              isDryRun: Boolean)(implicit ev: ApplicativeAsk[F, TraceId]): F[Option[Runtime]] =
         for {
           clusterOpt <- dependencies.dataprocService
             .getCluster(runtime.googleProject, regionName, DataprocClusterName(runtime.runtimeName))
-          _ <- clusterOpt.traverse_ { _ =>
-            if (isDryRun)
-              logger.warn(s"${runtime} still exists in Google. It needs to be deleted")
-            else
-              logger.warn(s"${runtime} still exists in Google. Going to delete") >> dependencies.dataprocService
-                .deleteCluster(runtime.googleProject, regionName, DataprocClusterName(runtime.runtimeName))
-                .void
+          _ <- clusterOpt.traverse_ { cluster =>
+            if (cluster.getStatus.getState.name() == "Error")
+              logger.warn(s"${runtime} still exists in Google in Error state. User might want to delete the runtime")
+            else {
+              if (isDryRun)
+                logger.warn(
+                  s"${runtime} still exists in ${cluster.getStatus.getState.name()} status. It needs to be deleted"
+                )
+              else
+                logger.warn(s"${runtime} still exists in ${cluster.getStatus.getState.name()} status. Going to delete") >> dependencies.dataprocService
+                  .deleteCluster(runtime.googleProject, regionName, DataprocClusterName(runtime.runtimeName))
+                  .void
+            }
           }
         } yield clusterOpt.fold(none[Runtime])(_ => Some(runtime))
 
@@ -47,11 +50,11 @@ object DeletedRuntimeChecker {
         for {
           runtimeOpt <- dependencies.computeService
             .getInstance(runtime.googleProject, zoneName, InstanceName(runtime.runtimeName))
-          _ <- runtimeOpt.traverse_ { _ =>
+          _ <- runtimeOpt.traverse_ { rt =>
             if (isDryRun)
-              logger.warn(s"${runtime} still exists in Google. It needs to be deleted")
+              logger.warn(s"${runtime} still exists in ${rt.getStatus} status. It needs to be deleted")
             else
-              logger.warn(s"${runtime} still exists in Google. Going to delete") >>
+              logger.warn(s"${runtime} still exists in ${rt.getStatus} status. Going to delete") >>
                 dependencies.computeService
                   .deleteInstance(runtime.googleProject, zoneName, InstanceName(runtime.runtimeName))
                   .void
